Add xarray plotting support and related workflow updates
