+++ src/utils.py
@@ -4,175 +4,8 @@ import plotly.graph_objects as go
 import plotly.express as px
 import os
 from plotly.subplots import make_subplots
-from sklearn.decomposition import FastICA
-import scipy.signal as signal   
 import pandas as pd
-import xarray as xr
 
-
-
-
-
-### PLOTS ####
-
-def plot_xarray_with_regions(xr_dataarray, regions):
-    """
-    Plot a DataArray with events highlighted as background colors.
-
-    Parameters
-    ----------
-    xr_dataarray : xarray.DataArray
-        The data array to plot. The function accepts either a 1D signal or a
-        2D array with a ``time`` dimension and an optional ``channel`` dimension.
-    regions : list of dict
-        List of dictionaries, each containing 'span' (tuple of start_time, end_time)
-        and 'name' (str) for each event region.
-
-    Returns
-    -------
-    None
-        Displays the plot using matplotlib.
-    """
-    if "time" not in xr_dataarray.coords:
-        raise ValueError("xr_dataarray must have a 'time' coordinate.")
-
-    time = np.asarray(xr_dataarray.coords["time"].values, dtype=float)
-    values = np.asarray(xr_dataarray.values)
-
-    if values.ndim == 1:
-        series_list = [values]
-        labels = ["Signal"]
-    elif values.ndim == 2:
-        if values.shape[0] == len(time):
-            series_list = [values[:, idx] for idx in range(values.shape[1])]
-            if "channel" in xr_dataarray.coords:
-                labels = [str(ch) for ch in xr_dataarray.coords["channel"].values]
-            else:
-                labels = [f"Signal {idx + 1}" for idx in range(values.shape[1])]
-        elif values.shape[1] == len(time):
-            series_list = [values[idx, :] for idx in range(values.shape[0])]
-            if "channel" in xr_dataarray.coords:
-                labels = [str(ch) for ch in xr_dataarray.coords["channel"].values]
-            else:
-                labels = [f"Signal {idx + 1}" for idx in range(values.shape[0])]
-        else:
-            raise ValueError("Unable to align 2D data with the time coordinate.")
-    else:
-        raise ValueError("Expected a 1D or 2D DataArray for plotting.")
-
-    plt.figure(figsize=(12, 6))
-    for series, label in zip(series_list, labels):
-        plt.plot(time, series, label=label)
-
-    for region in regions:
-        start, end = region["span"]
-        name = region["name"]
-        plt.axvspan(start, end, alpha=0.3, label=name)
-
-    plt.xlabel("Time (s)")
-    plt.ylabel("Signal")
-    plt.legend()
-    plt.show()
-
-def plot_signal_with_events(time, data, channels, marker_channel, event_to_marker, selected_time):
-    """
-    Plot signal data with background colors indicating different events.
-    
-    This function creates a time-series plot with colored background regions corresponding
-    to different events. Each event type is assigned a unique color and plotted as a 
-    semi-transparent background. Multiple signal channels can be overlaid on the same plot.
-    
-    Parameters
-    ----------
-    time : array-like
-        Time vector in seconds, matching the length of data columns
-    data : numpy.ndarray
-        Signal data array with shape (n_channels, n_samples)
-    channels : list of str
-        List of channel names corresponding to rows in data
-    marker_channel : array-like
-        Array of marker values indicating which event is active at each time point.
-        0 indicates no event, other values map to events via event_to_marker dict.
-        Should have same length as time vector.
-    event_to_marker : dict
-        Dictionary mapping event names (str) to marker values (int).
-        Example: {'Peppa': 1, 'Incredibles': 2, 'Brave': 3}
-    selected_time : list of float
-        [start_time, end_time] in seconds, used for plot title
-    
-    Returns
-    -------
-    None
-        Displays the plot using matplotlib
-    
-    Notes
-    -----
-    - Event regions are plotted as semi-transparent (alpha=0.3) background spans
-    - Up to 6 default colors are cycled for different events
-    - Each event is only added to the legend once, even if it appears multiple times
-    - Grid is enabled with alpha=0.3 for better readability
-    
-    Examples
-    --------
-    >>> time, channels, data = multimodal_data.get_signals(
-    ...     mode='EEG', member='ch', 
-    ...     selected_channels=['Fz', 'Cz'], 
-    ...     selected_times=[60, 120]
-    ... )
-    >>> time, marker_channel, event_to_marker = multimodal_data.get_events_as_marker_channel(
-    ...     selected_times=[60, 120]
-    ... )
-    >>> plot_signal_with_events(time, data, channels, marker_channel, 
-    ...                          event_to_marker, [60, 120])
-    """
-    # Create reverse mapping from marker value to event name
-    marker_to_event = {v: k for k, v in event_to_marker.items()}
-
-    # Create color map for events
-    colors = ['lightblue', 'lightgreen', 'lightyellow', 'lightcoral', 'lightpink', 'lavender']
-    event_colors = {}
-    for i, event_name in enumerate(marker_to_event.values()):
-        if event_name != '':
-            event_colors[event_name] = colors[i % len(colors)]
-
-    plt.figure(figsize=(15, 6), dpi=100)
-
-    # Plot background colors for each event
-    current_marker = marker_channel[0]
-    segment_start = time[0]
-
-    for i in range(1, len(marker_channel)):
-        if marker_channel[i] != current_marker:
-            # End of current segment
-            if current_marker > 0 and current_marker in marker_to_event:
-                event_name = marker_to_event[current_marker]
-                if event_name in event_colors:
-                    plt.axvspan(segment_start, time[i-1], 
-                            alpha=0.6, color=event_colors[event_name], 
-                            label=event_name if segment_start == time[0] or event_name not in plt.gca().get_legend_handles_labels()[1] else '')
-            # Start new segment
-            current_marker = marker_channel[i]
-            segment_start = time[i]
-
-    # Handle last segment
-    if current_marker > 0 and current_marker in marker_to_event:
-        event_name = marker_to_event[current_marker]
-        if event_name in event_colors:
-            plt.axvspan(segment_start, time[-1], 
-                    alpha=0.6, color=event_colors[event_name],
-                    label=event_name if event_name not in plt.gca().get_legend_handles_labels()[1] else '')
-
-    # Plot  data
-    for idxch, ch in enumerate(channels):
-            plt.plot(time, data[:, idxch], label=ch, linewidth=1.0)
-
-    plt.xlabel('Time (s)')
-    plt.ylabel('Amplitude (µV)')
-    plt.title(f'EEG Signal ({", ".join(channels)}) and Event Markers between {selected_time[0]}s and {selected_time[1]}s')
-    plt.legend(loc='upper right')
-    plt.grid(True, alpha=0.3)
-    plt.tight_layout()
-    plt.show()
     
 
 def plot_eeg_channels_pl(mmd, selected_events, selected_channels, title='Filtered EEG Channels', renderer='auto'):
@@ -434,218 +267,7 @@ def overlay_eeg_channels_hyperscanning_pl(data_ch, data_cg,
             save_figure_to_html(fig, title, event)
 
 
-# ==================================
 
-def plot_filter_characteristics(b, a, f, T, Fs, f_lim=None, db_lim=None):
-    """
-    Plot comprehensive filter characteristics including magnitude response, 
-    group delay, impulse response, and step response.
-    
-    Parameters
-    ----------
-    b : array_like
-        Numerator polynomial coefficients of the filter
-    a : array_like
-        Denominator polynomial coefficients of the filter
-    f : array_like
-        Frequency vector for frequency response computation (Hz)
-    T : float
-        Time duration for impulse and step response plots (seconds)
-    Fs : float
-        Sampling frequency (Hz)
-    f_lim : tuple of float, optional
-        Frequency axis limits (f_min, f_max) in Hz. Default is (0, Fs/2)
-    db_lim : tuple of float, optional
-        Magnitude axis limits (db_min, db_max) in dB. Default is auto-scaled
-    
-    Returns
-    -------
-    fig : matplotlib.figure.Figure
-        The created figure object
-    """
-    # Set default frequency limits
-    if f_lim is None:
-        f_lim = (0, Fs / 2)
-    else:
-        f_lim = (max(f_lim[0], 0), min(f_lim[1], Fs / 2))   
-        f = f[np.logical_and(f >= f_lim[0], f <= f_lim[1])]
-    
-    # Create time vector
-    time_vector = np.arange(-T, T, 1 / Fs)
-
-    # Compute frequency response
-    frequencies, frequency_response = signal.freqz(b, a, f, fs=Fs)
-    magnitude = np.abs(frequency_response)
-    magnitude_db = 20 * np.log10(magnitude)
-
-    # Compute group delay
-    _, group_delay = signal.group_delay((b, a), f, fs=Fs)
-
-    # Compute impulse response
-    impulse_signal = np.zeros(len(time_vector))
-    impulse_signal[len(time_vector) // 2] = 1
-    impulse_response = signal.lfilter(b, a, impulse_signal)
-
-    # Compute step response
-    step_signal = np.zeros(len(time_vector))
-    step_signal[len(time_vector) // 2:] = 1
-    step_response = signal.lfilter(b, a, step_signal)
-
-    # Create figure with subplots
-    fig = plt.figure(figsize=(15, 10))
-    
-    # Plot magnitude response (top left)
-    _plot_magnitude_response(frequencies, magnitude_db, f_lim, db_lim)
-    
-    # Plot group delay (bottom left)
-    _plot_group_delay(frequencies, group_delay, f_lim)
-    
-    # Plot impulse response (top right)
-    _plot_impulse_response(time_vector, impulse_signal, impulse_response, T)
-    
-    # Plot step response (bottom right)
-    _plot_step_response(time_vector, step_signal, step_response, T)
-    
-    fig.subplots_adjust(hspace=0.5)
-    plt.show()
-    
-    return fig
-
-
-def _find_db_crossings(frequencies, magnitude_db, target_db):
-    """
-    Find frequencies where magnitude response crosses a target dB value.
-    
-    Parameters
-    ----------
-    frequencies : array_like
-        Frequency vector (Hz)
-    magnitude_db : array_like
-        Magnitude response in dB
-    target_db : float
-        Target dB value to find crossings
-    
-    Returns
-    -------
-    list
-        List of frequencies where crossings occur
-    """
-    crossings = []
-    for i in range(len(magnitude_db) - 1):
-        # Check if the line segment crosses the target
-        if (magnitude_db[i] >= target_db and magnitude_db[i + 1] < target_db) or \
-           (magnitude_db[i] < target_db and magnitude_db[i + 1] >= target_db):
-            # Linear interpolation to find exact crossing frequency
-            freq_crossing = frequencies[i] + (target_db - magnitude_db[i]) * \
-                           (frequencies[i + 1] - frequencies[i]) / \
-                           (magnitude_db[i + 1] - magnitude_db[i])
-            crossings.append(freq_crossing)
-    return crossings
-
-
-def _plot_magnitude_response(frequencies, magnitude_db, f_lim, db_lim):
-    """
-    Plot magnitude response of the filter in dB.
-    
-    Parameters
-    ----------
-    frequencies : array_like
-        Frequency vector (Hz)
-    magnitude_db : array_like
-        Magnitude response in dB
-    f_lim : tuple of float
-        Frequency axis limits (f_min, f_max)
-    db_lim : tuple of float or None
-        Magnitude axis limits (db_min, db_max)
-    """
-    plt.subplot(2, 2, 1)
-    plt.title('Magnitude Response')
-    plt.plot(frequencies, magnitude_db)
-    plt.ylabel('Magnitude [dB]')
-    plt.grid(True)
-    plt.xlim(f_lim)
-    
-    # Find and mark -6dB crossing points
-    crossing_freqs = _find_db_crossings(frequencies, magnitude_db, -6.0)
-    for freq in crossing_freqs:
-        plt.axvline(x=freq, color='r', linestyle='--', linewidth=1.5, alpha=0.7)
-    
-    if db_lim is None:
-        # Auto-scale to show only the zoomed frequency range
-        magnitude_zoom = magnitude_db[np.logical_and(f_lim[0] < frequencies, frequencies < f_lim[1])]
-        plt.ylim((np.min(magnitude_zoom), np.max(magnitude_zoom)))
-    else:
-        plt.ylim((db_lim[0], db_lim[1]))
-
-
-def _plot_group_delay(frequencies, group_delay, f_lim):
-    """
-    Plot group delay of the filter.
-    
-    Parameters
-    ----------
-    frequencies : array_like
-        Frequency vector (Hz)
-    group_delay : array_like
-        Group delay in samples
-    f_lim : tuple of float
-        Frequency axis limits (f_min, f_max)
-    """
-    plt.subplot(2, 2, 3)
-    plt.title('Group Delay')
-    plt.plot(frequencies, group_delay)
-    plt.ylabel('Samples')
-    plt.xlabel('Frequency [Hz]')
-    plt.grid(True)
-    plt.xlim(f_lim)
-    plt.ylim([np.min(group_delay) - 1, np.max(group_delay) + 1])
-
-
-def _plot_impulse_response(time_vector, impulse_signal, impulse_response, T):
-    """
-    Plot impulse response of the filter.
-    
-    Parameters
-    ----------
-    time_vector : array_like
-        Time vector (seconds)
-    impulse_signal : array_like
-        Input impulse signal
-    impulse_response : array_like
-        Filter's impulse response
-    T : float
-        Time duration (seconds)
-    """
-    plt.subplot(2, 2, 2)
-    plt.title('Impulse Response')
-    plt.stem(time_vector, impulse_signal)
-    plt.plot(time_vector, impulse_response, 'r')
-    plt.xlim([-T / 4, T])
-    plt.grid(True)
-
-
-def _plot_step_response(time_vector, step_signal, step_response, T):
-    """
-    Plot step response of the filter.
-    
-    Parameters
-    ----------
-    time_vector : array_like
-        Time vector (seconds)
-    step_signal : array_like
-        Input step signal
-    step_response : array_like
-        Filter's step response
-    T : float
-        Time duration (seconds)
-    """
-    plt.subplot(2, 2, 4)
-    plt.title('Step Response')
-    plt.plot(time_vector, step_signal)
-    plt.plot(time_vector, step_response, 'r')
-    plt.xlim([-T / 4, T])
-    plt.xlabel('Time [s]')
-    plt.grid(True)
 
 # ==================================
